refactor: Top3Image.tsx を分割する (#70)
