+++ src/hooks/useImageCapture.ts
@@ -1,0 +1,86 @@
+import { useRef, useState, useCallback, useEffect } from 'react'
+import html2canvas from 'html2canvas'
+import {
+  sanitizeFilename,
+  getImageErrorMessage,
+  formatDate,
+} from '../utils/image-helpers'
+
+const IMAGE_SIZE = 1080
+const CANVAS_BG = '#111827'
+
+export function useImageCapture(theme: string) {
+  const captureRef = useRef<HTMLDivElement>(null)
+  const containerRef = useRef<HTMLDivElement>(null)
+  const [isGenerating, setIsGenerating] = useState(false)
+  const [error, setError] = useState<string | null>(null)
+  const [successOpen, setSuccessOpen] = useState(false)
+  const [scale, setScale] = useState(0.5)
+
+  useEffect(() => {
+    const container = containerRef.current
+    if (!container) return
+
+    const updateScale = () => {
+      const containerWidth = container.clientWidth
+      const newScale =
+        Math.round(Math.min(containerWidth / IMAGE_SIZE, 0.5) * 1000) / 1000
+      setScale(newScale)
+    }
+
+    updateScale()
+
+    const observer = new ResizeObserver(updateScale)
+    observer.observe(container)
+    return () => observer.disconnect()
+  }, [])
+
+  const handleDownload = useCallback(async () => {
+    if (!captureRef.current) return
+
+    setIsGenerating(true)
+    setError(null)
+    try {
+      const canvas = await html2canvas(captureRef.current, {
+        width: IMAGE_SIZE,
+        height: IMAGE_SIZE,
+        scale: 1,
+        useCORS: true,
+        allowTaint: false,
+        backgroundColor: CANVAS_BG,
+      })
+
+      const dataUrl = canvas.toDataURL('image/png')
+      if (!dataUrl || dataUrl === 'data:,') {
+        setError('画像の生成に失敗しました。画像データが空です。')
+        return
+      }
+
+      const date = formatDate()
+      const safeTheme = theme ? `-${sanitizeFilename(theme)}` : ''
+      const link = document.createElement('a')
+      link.download = `my-no1s-${date}${safeTheme}.png`
+      link.href = dataUrl
+      link.click()
+      setSuccessOpen(true)
+    } catch (err) {
+      console.error('[Top3Image] Failed to generate image:', err)
+      setError(getImageErrorMessage(err))
+    } finally {
+      setIsGenerating(false)
+    }
+  }, [theme])
+
+  return {
+    captureRef,
+    containerRef,
+    isGenerating,
+    error,
+    setError,
+    successOpen,
+    setSuccessOpen,
+    scale,
+    handleDownload,
+    IMAGE_SIZE,
+  }
+}
